docker_exec:
The SIGTERM was caught by PID 1 but once it killed the child_for_exec
process, the main code wasn't executed anymore and in a racy fashion,
the end of the _term function wasn't completed.

Considering the _term function would exit and return to the main was a big mistake.
Once the _term is engaged, all the exit/cleanup code have to be run in this function.

The patch is then moving all the end of exec() function into the _term()
function. As all containers will be stopped by a SIGTERM, this code will
always be executed.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1624341

Signed-off-by: Erwan Velu <erwan@redhat.com>